% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{career_profile}
\alias{career_profile}
\alias{hero_profile}
\title{Profile API}
\source{
https://dev.battle.net/io-docs
}
\usage{
career_profile(battletag)

hero_profile(battletag, id)
}
\arguments{
\item{battletag}{Battle Tag in name-#### format (ie. Noob-1234)}

\item{id}{The hero id of the hero to look up}
}
\description{
The Character Profile API is the primary way to access character information.
}
\details{
\describe{
 \item{career_profile}{Returns the career profile of a Battle Tag}
 \item{hero_profile}{Returns the hero profile of a Battle Tag's hero}
}
}

