% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_resources.R
\name{item}
\alias{artisan}
\alias{follower}
\alias{item}
\title{Data Resources API}
\source{
https://dev.battle.net/io-docs
}
\usage{
item(data)

follower(follower = c("enchantress", "scoundrel", "templar"))

artisan(artisan = c("blacksmith", "jeweler", "mystic"))
}
\arguments{
\item{data}{The item data string (from a profile) containing the item to lookup}

\item{follower}{The data about a follower}

\item{artistan}{The data about an artisan}
}
\details{
The Data Resources API is the primary way to access character information.
\describe{
 \item{item}{Returns data for a profile item}
 \item{follower}{Returns data for a follower}
 \item{artisan}{Returns data for an artisan}
}
}

